--- Gestion-User/views/Interface_Camper.fxml
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.image.Image?>
-<?import javafx.scene.image.ImageView?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.text.Font?>
-
-<AnchorPane id="AnchorPane" prefHeight="630.0" prefWidth="1043.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="edu.SprintJava2.views.Interface_CamperController">
-   <children>
-      <AnchorPane prefHeight="630.0" prefWidth="200.0" style="-fx-background-color: ADC47F;">
-         <children>
-            <Button layoutX="25.0" layoutY="115.0" mnemonicParsing="false" onAction="#profile" prefHeight="31.0" prefWidth="151.0" text="My Profile" />
-            <Button layoutX="25.0" layoutY="187.0" mnemonicParsing="false" onAction="#contact" prefHeight="31.0" prefWidth="151.0" text="Contact Us " />
-         </children>
-      </AnchorPane>
-      <Button layoutX="962.0" layoutY="23.0" mnemonicParsing="false" onAction="#retourner" text="Log out" />
-      <Label layoutX="244.0" layoutY="44.0" prefHeight="50.0" prefWidth="327.0" text="Welcome Camper">
-         <font>
-            <Font name="System Bold" size="34.0" />
-         </font>
-      </Label>
-      <ImageView fitHeight="113.0" fitWidth="127.0" layoutX="560.0" layoutY="13.0" pickOnBounds="true" preserveRatio="true">
-         <image>
-            <Image url="file:/C:/Users/chayma/Downloads/camper-van.png" />
-         </image>
-      </ImageView>
-   </children>
-</AnchorPane>
