--- Gestion-User/views/Interface_Host.fxml
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.image.Image?>
-<?import javafx.scene.image.ImageView?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.text.Font?>
-
-<AnchorPane id="AnchorPane" prefHeight="630.0" prefWidth="1043.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="edu.SprintJava2.views.Interface_HostController">
-   <children>
-      <AnchorPane layoutX="-3.0" prefHeight="630.0" prefWidth="189.0" style="-fx-background-color: ADC47F;">
-         <children>
-            <Button layoutX="56.0" layoutY="145.0" mnemonicParsing="false" onAction="#profil" text="My Profile" />
-            <Button layoutX="56.0" layoutY="213.0" mnemonicParsing="false" onAction="#contact" prefHeight="31.0" prefWidth="88.0" text="Conatc Us" />
-         </children>
-      </AnchorPane>
-      <Label layoutX="278.0" layoutY="69.0" prefHeight="56.0" prefWidth="415.0" text="Welcome Host">
-         <font>
-            <Font name="System Bold" size="39.0" />
-         </font>
-      </Label>
-      <ImageView fitHeight="150.0" fitWidth="200.0" layoutX="579.0" layoutY="39.0" pickOnBounds="true" preserveRatio="true" />
-      <ImageView fitHeight="110.0" fitWidth="103.0" layoutX="562.0" layoutY="30.0" pickOnBounds="true" preserveRatio="true">
-         <image>
-            <Image url="file:/C:/Users/chayma/Downloads/tour-guide.png" />
-         </image>
-      </ImageView>
-      <Button layoutX="905.0" layoutY="39.0" mnemonicParsing="false" onAction="#retourner" text="Log out" />
-   </children>
-</AnchorPane>
