--- Gestion-User/views/Interface_Provider.fxml
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.image.Image?>
-<?import javafx.scene.image.ImageView?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.text.Font?>
-
-<AnchorPane id="AnchorPane" prefHeight="630.0" prefWidth="1043.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="edu.SprintJava2.views.Interface_ProviderController">
-   <children>
-      <AnchorPane prefHeight="630.0" prefWidth="221.0" style="-fx-background-color: ADC47F;">
-         <children>
-            <Button layoutX="37.0" layoutY="96.0" mnemonicParsing="false" onAction="#profile1" prefHeight="31.0" prefWidth="147.0" text="My profile" />
-            <Button layoutX="37.0" layoutY="161.0" mnemonicParsing="false" onAction="#contact1" prefHeight="31.0" prefWidth="147.0" text="Contact Us " />
-         </children>
-      </AnchorPane>
-      <Button layoutX="944.0" layoutY="24.0" mnemonicParsing="false" onAction="#logout" text="Log out" />
-      <ImageView fitHeight="83.0" fitWidth="147.0" layoutX="671.0" layoutY="40.0" pickOnBounds="true" preserveRatio="true" />
-      <Label layoutX="260.0" layoutY="40.0" prefHeight="47.0" prefWidth="359.0" text="Welcome Provider">
-         <font>
-            <Font name="System Bold" size="32.0" />
-         </font>
-      </Label>
-      <ImageView fitHeight="114.0" fitWidth="175.0" layoutX="581.0" layoutY="25.0" pickOnBounds="true" preserveRatio="true">
-         <image>
-            <Image url="file:/C:/Users/chayma/Downloads/service.png" />
-         </image>
-      </ImageView>
-   </children>
-</AnchorPane>
