+++ Gestion-User/views/Interface_User.fxml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.layout.AnchorPane?>
+
+<AnchorPane id="AnchorPane" prefHeight="568.0" prefWidth="827.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="edu.SprintJava2.views.Interface_UserController">
+   <children>
+      <AnchorPane prefHeight="568.0" prefWidth="124.0" style="-fx-background-color: ADC47F;">
+         <children>
+            <Button layoutX="18.0" layoutY="103.0" mnemonicParsing="false" onAction="#profile" text="My Profile" />
+         </children>
+      </AnchorPane>
+      <Label fx:id="role" layoutX="147.0" layoutY="31.0" prefHeight="64.0" prefWidth="288.0" text="Label" />
+      <Button layoutX="726.0" layoutY="31.0" mnemonicParsing="false" onAction="#logout" text="Log out " />
+   </children>
+</AnchorPane>
