+++ Gestion-User/views/Interface_UserController.java
@@ -1,0 +1,57 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package edu.SprintJava2.views;
+
+import edu.SprintJava2.services.UsersSession;
+import java.net.URL;
+import java.util.ResourceBundle;
+import javafx.event.ActionEvent;
+import javafx.fxml.FXML;
+import javafx.fxml.Initializable;
+import javafx.scene.control.Alert;
+import javafx.scene.control.ButtonType;
+import javafx.scene.control.Label;
+import javafx.stage.Stage;
+
+/**
+ * FXML Controller class
+ *
+ * @author chayma
+ */
+public class Interface_UserController implements Initializable {
+
+    @FXML
+    private Label role;
+
+    /**
+     * Initializes the controller class.
+     */
+    @Override
+    public void initialize(URL url, ResourceBundle rb) {
+        // TODO
+    }    
+
+    @FXML
+    private void profile(ActionEvent event) {
+         NewFXMain.setScene("Userprofile");
+    }
+
+
+    @FXML
+    private void logout(ActionEvent event) {
+        
+        Stage stage;
+        Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+        alert.setTitle("Logout");
+        alert.setHeaderText("You're about to logout");
+        alert.setContentText("Do you want to logout "+UsersSession.getName()+"?");
+        if (alert.showAndWait().get() == ButtonType.OK ) {
+       NewFXMain.setScene("InterfaceLogin");
+    }
+        
+    }
+    }
+
